feat: restructure category module, add new routes and services, and update Jest configuration
