+++ src/modules/category/__tests__/model.spec.ts
@@ -1,0 +1,5 @@
+describe('Category model', () => {
+    it('should be defined', () => {
+        expect(true).toBe(true);
+    });
+});
