+++ src/modules/category/__tests__/routes.spec.ts
@@ -1,0 +1,5 @@
+describe('Category routes', () => {
+    it('should be defined', () => {
+        expect(true).toBe(true);
+    });
+});
