+++ src/modules/category/__tests__/services.spec.ts
@@ -1,0 +1,5 @@
+describe('Category services', () => {
+    it('should be defined', () => {
+        expect(true).toBe(true);
+    });
+});
